refactor(default-theme): segregation of checkout components
