+++ packages/default-theme/components/checkout/steps/PaymentStep.vue
@@ -4,104 +4,8 @@
       title="3. Payment"
       class="sf-heading--left sf-heading--no-underline title"
     />
-    <div class="form">
-      <SfCheckbox
-        v-model="differentThanShipping"
-        label="Use different address for billing"
-        name="copyShippingAddress"
-        class="form__element"
-      />
-    </div>
-    <div class="form" v-if="differentThanShipping">
-      <SfInput
-        v-model="firstName"
-        :valid="!validations.firstName.$error"
-        error-message="This field is required"
-        label="First name"
-        name="firstName"
-        class="form__element form__element--half"
-        required
-      />
-      <SfInput
-        v-model="lastName"
-        :valid="!validations.lastName.$error"
-        error-message="This field is required"
-        label="Last name"
-        name="lastName"
-        class="form__element form__element--half form__element--half-even"
-        required
-      />
-      <SfInput
-        v-model="street"
-        :valid="!validations.street.$error"
-        error-message="This field is required"
-        label="Street name"
-        name="street"
-        class="form__element"
-        required
-      />
-      <SfInput
-        v-model="apartment"
-        :valid="!validations.apartment.$error"
-        error-message="This field is required"
-        label="House/Apartment number"
-        name="apartment"
-        class="form__element"
-        required
-      />
-      <SfInput
-        v-model="city"
-        :valid="!validations.city.$error"
-        error-message="This field is required"
-        label="City"
-        name="city"
-        class="form__element form__element--half"
-        required
-      />
-      <SfInput
-        v-model="state"
-        :valid="!validations.state.$error"
-        error-message="This field is required"
-        label="State/Province"
-        name="state"
-        class="form__element form__element--half form__element--half-even"
-        required
-      />
-      <SfInput
-        v-model="zipcode"
-        :valid="!validations.zipcode.$error"
-        error-message="This field is required"
-        label="Zip-code"
-        name="zipcode"
-        class="form__element form__element--half"
-        required
-      />
-      <SfSelect
-        v-model="countryId"
-        :valid="!validations.countryId.$error"
-        error-message="This field is required"
-        label="Country"
-        class="form__element form__element--half form__element--half-even form__select sf-select--underlined"
-        required
-      >
-        <SfSelectOption
-          v-for="countryOption in countries"
-          :key="countryOption"
-          :value="countryOption"
-        >
-          {{ countryOption }}
-        </SfSelectOption>
-      </SfSelect>
-      <SfInput
-        v-model="phoneNumber"
-        :valid="!validations.phoneNumber.$error"
-        error-message="This field is required"
-        label="Phone number"
-        name="phone"
-        class="form__element"
-        required
-      />
-    </div>
+    <BillingAddressGuestForm v-if="isGuestOrder" />
+    <BillingAddressUserForm v-else />
     <SfHeading
       title="Payment methods"
       subtitle="Choose your payment method"
@@ -229,27 +133,25 @@ import {
   SfHeading,
   SfInput,
   SfButton,
-  SfSelect,
   SfRadio,
   SfImage,
   SfCheckbox,
 } from '@storefront-ui/vue'
-import { validationMixin } from 'vuelidate'
-import {
-  usePaymentStep,
-  usePaymentStepValidationRules,
-} from '@shopware-pwa/default-theme/logic/checkout/usePaymentStep'
+import BillingAddressGuestForm from '@shopware-pwa/default-theme/components/checkout/steps/guest/BillingAddressGuestForm'
+import BillingAddressUserForm from '@shopware-pwa/default-theme/components/checkout/steps/user/BillingAddressUserForm'
+import { useCheckout } from '@shopware-pwa/composables'
+
 export default {
-  name: 'Payment',
-  mixins: [validationMixin],
+  name: 'PaymentStep',
   components: {
     SfHeading,
     SfInput,
     SfButton,
-    SfSelect,
     SfRadio,
     SfImage,
     SfCheckbox,
+    BillingAddressGuestForm,
+    BillingAddressUserForm,
   },
   data() {
     return {
@@ -265,46 +167,9 @@ export default {
     }
   },
   setup() {
-    const {
-      validations,
-      setValidations,
-      validate,
-      firstName,
-      lastName,
-      street,
-      apartment,
-      city,
-      state,
-      zipcode,
-      countryId,
-      phoneNumber,
-      differentThanShipping,
-    } = usePaymentStep()
-    return {
-      validations,
-      setValidations,
-      firstName,
-      lastName,
-      street,
-      apartment,
-      city,
-      state,
-      zipcode,
-      countryId,
-      phoneNumber,
-      differentThanShipping,
-    }
-  },
-  watch: {
-    $v: {
-      immediate: true,
-      handler: function () {
-        this.setValidations(this.$v)
-      },
-    },
-  },
-  validations: {
-    ...usePaymentStepValidationRules,
+    const { isGuestOrder } = useCheckout()
+
+    return { isGuestOrder }
   },
   computed: {
     isCreditCard() {
