+++ packages/default-theme/components/checkout/summary/OrderItem.vue
