+++ packages/default-theme/components/checkout/summary/OrderItemsTable.vue
